Create harshPr4
Get the input of an array of n' data elements Add a string to false at first. For each length ‘i’ call the Combination() function In Combination(), if currLen is more than the reqLen then return Otherwise, if currLen is equal to reqLen then there will be a new sequence generate Or proceed with a start as ‘false’ and recursively call Combination() with incremented value of ‘start’ then Exit.